Move inline hooks to utils
